# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pompote-view API",
    "description": "Read-only HTTP surface for monitoring trading strategies. Every `/api/v1/*` route requires the `X-API-Key` header.",
    "contact": {
      "name": "Pierre Attard"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/monitoring/candles": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/candles`.",
        "operationId": "get_candles",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "description": "Exchange identifier as stored in `candles_5s.exchange`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "description": "Symbol identifier as stored in `candles_5s.symbol`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Aggregation width (must be one of [`Timeframe::ALLOWED`]).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `open_time` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `open_time` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated OHLCV buckets ordered by ascending `ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeframe, range or too many points.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error (e.g. schema drift).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies`.",
        "operationId": "list_strategies",
        "responses": {
          "200": {
            "description": "Every strategy, ordered by name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies/{id}/decisions": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies/{id}/decisions`.",
        "operationId": "get_strategy_decisions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strategy identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `executed_at` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `executed_at` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to [`MAX_ORDER_ROWS`], rejected above it).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live decisions (with market context) for the strategy on the window, ordered by ascending `created_at`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LiveDecisionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies/{id}/fills": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies/{id}/fills`.",
        "operationId": "get_strategy_fills",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strategy identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `executed_at` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `executed_at` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to [`MAX_ORDER_ROWS`], rejected above it).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live fills for the strategy on the window, ordered by ascending `executed_at`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LiveFillDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/monitoring/timeframes": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/timeframes`.",
        "operationId": "get_timeframes",
        "responses": {
          "200": {
            "description": "Allowed aggregation timeframes (finest to coarsest).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CandleDto": {
        "type": "object",
        "description": "HTTP DTO returned in the JSON array.\n\nWe deliberately down-cast `Decimal` → `f64` at the serialisation boundary:\nLightweight Charts requires JS numbers, and the relative error introduced\n(worst case ≈ 2.2e-16 for typical OHLC magnitudes) is invisible at chart\npixel resolution. The full-precision `Decimal` stays in the domain and the\npersistence layer — only this struct loses precision, and it is the only\nplace that does.",
        "required": [
          "ts",
          "o",
          "h",
          "l",
          "c",
          "v"
        ],
        "properties": {
          "c": {
            "type": "number",
            "format": "double",
            "description": "Close price."
          },
          "h": {
            "type": "number",
            "format": "double",
            "description": "High price."
          },
          "l": {
            "type": "number",
            "format": "double",
            "description": "Low price."
          },
          "o": {
            "type": "number",
            "format": "double",
            "description": "Open price."
          },
          "ts": {
            "type": "string",
            "format": "date-time",
            "description": "Bucket start, RFC3339 (e.g. `2026-05-27T14:00:00Z`)."
          },
          "v": {
            "type": "number",
            "format": "double",
            "description": "Total traded volume in the bucket."
          }
        }
      },
      "CandleErrorBody": {
        "type": "object",
        "description": "JSON error body returned on 4xx / 5xx by the candles endpoint.",
        "required": [
          "error"
        ],
        "properties": {
          "allowed": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Whitelisted `timeframe` values, surfaced for `invalid_timeframe`."
          },
          "error": {
            "type": "string",
            "description": "Machine-readable error discriminator (e.g. `invalid_timeframe`)."
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Maximum bucket count accepted by the endpoint for `too_many_points`.",
            "minimum": 0
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable explanation (optional)."
          },
          "requested": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Echoes the caller-supplied bucket count for `too_many_points`.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "x_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}